# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Educational Content Generator API",
    "description": "AI-powered educational content generation system built with Strands Agents",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/input-storage/search": {
      "get": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "Search Stored Inputs",
        "description": "Search stored inputs using semantic search and filters",
        "operationId": "search_stored_inputs_api_v1_input_storage_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query text",
              "title": "Query"
            },
            "description": "Search query text"
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by content type",
              "title": "Content Type"
            },
            "description": "Filter by content type"
          },
          {
            "name": "generation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by generation ID",
              "title": "Generation Id"
            },
            "description": "Filter by generation ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredInput"
                  },
                  "title": "Response Search Stored Inputs Api V1 Input Storage Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "Search Stored Inputs Post",
        "description": "Search stored inputs using POST request (for complex queries)",
        "operationId": "search_stored_inputs_post_api_v1_input_storage_search_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredInput"
                  },
                  "title": "Response Search Stored Inputs Post Api V1 Input Storage Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/input-storage/list": {
      "get": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "List Stored Inputs",
        "description": "List stored inputs with optional filters and pagination",
        "operationId": "list_stored_inputs_api_v1_input_storage_list_get",
        "parameters": [
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by content type",
              "title": "Content Type"
            },
            "description": "Filter by content type"
          },
          {
            "name": "generation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by generation ID",
              "title": "Generation Id"
            },
            "description": "Filter by generation ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredInputMetadata"
                  },
                  "title": "Response List Stored Inputs Api V1 Input Storage List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "List Stored Inputs Post",
        "description": "List stored inputs using POST request (for complex filters)",
        "operationId": "list_stored_inputs_post_api_v1_input_storage_list_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredInputMetadata"
                  },
                  "title": "Response List Stored Inputs Post Api V1 Input Storage List Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/input-storage/{storage_id}": {
      "get": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "Get Stored Input",
        "description": "Retrieve a specific stored input by ID",
        "operationId": "get_stored_input_api_v1_input_storage__storage_id__get",
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredInput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "Delete Stored Input",
        "description": "Delete a stored input and its metadata",
        "operationId": "delete_stored_input_api_v1_input_storage__storage_id__delete",
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/input-storage/{storage_id}/history": {
      "get": {
        "tags": [
          "input-storage",
          "input-storage"
        ],
        "summary": "Get Input History",
        "description": "Get processing history for inputs with the same content hash",
        "operationId": "get_input_history_api_v1_input_storage__storage_id__history_get",
        "parameters": [
          {
            "name": "storage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storage Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InputHistoryEntry"
                  },
                  "title": "Response Get Input History Api V1 Input Storage  Storage Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentType": {
        "type": "string",
        "enum": [
          "text",
          "file",
          "url"
        ],
        "title": "ContentType",
        "description": "Content type enumeration"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InputHistoryEntry": {
        "properties": {
          "storage_id": {
            "type": "string",
            "title": "Storage Id",
            "description": "Storage identifier"
          },
          "stored_at": {
            "type": "string",
            "title": "Stored At",
            "description": "Storage timestamp"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id",
            "description": "Associated generation ID"
          },
          "validation_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Validation Result",
            "description": "Validation results"
          },
          "processing_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Processing Metadata",
            "description": "Processing metadata"
          }
        },
        "type": "object",
        "required": [
          "storage_id",
          "stored_at"
        ],
        "title": "InputHistoryEntry",
        "description": "Entry in input processing history"
      },
      "InputListRequest": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Filter by content type"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id",
            "description": "Filter by generation ID"
          },
          "limit": {
            "type": "integer",
            "maximum": 200,
            "minimum": 1,
            "title": "Limit",
            "description": "Maximum number of results",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "title": "Offset",
            "description": "Number of results to skip",
            "default": 0
          }
        },
        "type": "object",
        "title": "InputListRequest",
        "description": "Request model for listing stored inputs"
      },
      "InputSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Search query text"
          },
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Filter by content type"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id",
            "description": "Filter by generation ID"
          },
          "limit": {
            "type": "integer",
            "maximum": 100,
            "minimum": 1,
            "title": "Limit",
            "description": "Maximum number of results",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "InputSearchRequest",
        "description": "Request model for searching stored inputs"
      },
...
      "StoredInput": {
        "properties": {
          "storage_id": {
            "type": "string",
            "title": "Storage Id",
            "description": "Unique storage identifier"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Stored content"
          },
          "metadata": {
            "$ref": "#/components/schemas/StoredInputMetadata",
            "description": "Input metadata"
          },
          "similarity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity Score",
            "description": "Similarity score for search results"
          },
          "search_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Snippet",
            "description": "Content snippet for search results"
          }
        },
        "type": "object",
        "required": [
          "storage_id",
          "content",
          "metadata"
        ],
        "title": "StoredInput",
        "description": "Complete stored input with content and metadata"
      },
      "StoredInputMetadata": {
        "properties": {
          "storage_id": {
            "type": "string",
            "title": "Storage Id",
            "description": "Unique storage identifier"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "Type of content"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash",
            "description": "SHA-256 hash of content"
          },
          "stored_at": {
            "type": "string",
            "format": "date-time",
            "title": "Stored At",
            "description": "When the input was stored"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id",
            "description": "Associated generation ID"
          },
          "content_length": {
            "type": "integer",
            "title": "Content Length",
            "description": "Length of content in characters"
          },
          "validation_status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Status",
            "description": "Whether validation passed"
          },
          "processing_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Status",
            "description": "Processing status"
          }
        },
        "type": "object",
        "required": [
          "storage_id",
          "content_type",
          "content_hash",
          "stored_at",
          "content_length"
        ],
        "title": "StoredInputMetadata",
        "description": "Metadata for stored input"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}